Remove Enable, Stop classes. Move pass through apps to enum in List command class.
