+++ src/main/java/bio/terra/cli/command/App.java
@@ -1,11 +1,9 @@
 package bio.terra.cli.command;
 
-import bio.terra.cli.command.app.Enable;
 import bio.terra.cli.command.app.Execute;
 import bio.terra.cli.command.app.GetImage;
 import bio.terra.cli.command.app.List;
 import bio.terra.cli.command.app.SetImage;
-import bio.terra.cli.command.app.Stop;
 import picocli.CommandLine.Command;
 
 /**
@@ -15,12 +13,5 @@ import picocli.CommandLine.Command;
 @Command(
     name = "app",
     description = "Commands related to applications in the Terra workspace context.",
-    subcommands = {
-      List.class,
-      Enable.class,
-      Stop.class,
-      GetImage.class,
-      SetImage.class,
-      Execute.class
-    })
+    subcommands = {List.class, GetImage.class, SetImage.class, Execute.class})
 public class App {}
