+++ src/main/java/bio/terra/cli/command/app/List.java
@@ -1,6 +1,5 @@
 package bio.terra.cli.command.app;
 
-import bio.terra.cli.apps.interfaces.Supported;
 import java.util.concurrent.Callable;
 import picocli.CommandLine.Command;
 
@@ -12,10 +11,24 @@ public class List implements Callable<Integer> {
   public Integer call() {
     System.out.println(
         "Call any of the supported applications listed below, by prefixing it with 'terra' (e.g. terra gsutil ls, terra nextflow run hello)\n");
-    for (Supported.SupportedApp app : Supported.SupportedApp.values()) {
+    for (PassThrough app : PassThrough.values()) {
       System.out.println("  " + app);
     }
 
     return 0;
+  }
+
+  /**
+   * This enum specifies the list returned by the 'terra app list' command. These commands are
+   * hidden top-level commands that can be called in a pass-through manner, meaning the user can
+   * call them as if they were running the commands locally, except with a "terra" prefix. This
+   * prefix means that the command will run within the context of a Terra workspace (e.g. in the
+   * backing Google project, wtih data references scoped to the workspace, etc).
+   */
+  public enum PassThrough {
+    nextflow,
+    gcloud,
+    gsutil,
+    bq
   }
 }
