+++ src/main/java/bio/terra/cli/command/app/passthrough/Bq.java
@@ -1,6 +1,6 @@
-package bio.terra.cli.command.app.supported;
+package bio.terra.cli.command.app.passthrough;
 
-import bio.terra.cli.apps.NextflowRunner;
+import bio.terra.cli.apps.DockerAppsRunner;
 import bio.terra.cli.auth.AuthenticationManager;
 import bio.terra.cli.context.GlobalContext;
 import bio.terra.cli.context.WorkspaceContext;
@@ -9,12 +9,9 @@ import java.util.concurrent.Callable;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
-/** This class corresponds to the second-level "terra nextflow" command. */
-@Command(
-    name = "nextflow",
-    description = "Use the nextflow tool in the Terra workspace.",
-    hidden = true)
-public class Nextflow implements Callable<Integer> {
+/** This class corresponds to the second-level "terra bq" command. */
+@Command(name = "bq", description = "Use the bq tool in the Terra workspace.", hidden = true)
+public class Bq implements Callable<Integer> {
 
   @CommandLine.Unmatched private List<String> cmdArgs;
 
@@ -24,9 +21,12 @@ public class Nextflow implements Callable<Integer> {
     WorkspaceContext workspaceContext = WorkspaceContext.readFromFile();
 
     new AuthenticationManager(globalContext, workspaceContext).loginTerraUser();
-    NextflowRunner nextflowRunner = new NextflowRunner();
-    nextflowRunner.setContext(globalContext, workspaceContext);
-    String cmdOutput = nextflowRunner.run(cmdArgs);
+    String fullCommand = DockerAppsRunner.buildFullCommand("bq", cmdArgs);
+
+    // no need for any special setup or teardown logic since bq is already initialized when the
+    // container starts
+    String cmdOutput =
+        new DockerAppsRunner(globalContext, workspaceContext).runToolCommand(fullCommand);
     System.out.println(cmdOutput);
 
     return 0;
